Create GroupBy - Having - Limit
Here is an example of SQL code that is a little more complex.  It uses the GROUP BY, HAVING, and LIMIT clauses to filter the data.  This code can be useful for reference.  The "data" in this file is fictious and not meant to represent any project in any way.